document: nft 관리를 위한 nft default 컴포넌트 생성
